Move PrimitiveCellValue to a subfolder
